changed generation of codes and phone_number Type

+++ backAPI/src/main/java/com/task/DTO/Records/ClientsRecord.java
@@ -5,5 +5,5 @@ public record ClientsRecord(
         String name,
         String surname,
         String email,
-        Integer number
+        Long number
 ) {}
